pyproject.toml
[project]
name = "e-commerce-chatbot-with-product-recommendations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.97.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]
